requirements.txt
flask==3.0.2
werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.1
pytz==2024.1
python-facebook-api==0.15.0